Custom example of a DataTable using Flet and Python.
I felt the need to have a DataTable where I could reuse it in other logics and thus avoid rewriting. With the help of documentation and partners from the web, I already used DataTable in my projects through the Programadoe Aventureiro course. I felt the need to edit data in the table and, on the other hand, delete rows from the table. How could I do this if clicking on the column only marked the checkbox? So I created a column for the checkboxes. But how to implement the same behavior as the on_select_changed function? After a lot of struggle and with the help of two artificial intelligences, Chat GPT and BlackBox, we arrived at the example below: